split into workspace: termpair (client), termpair-server, termpair-common
