+++ install.sh
@@ -106,10 +106,10 @@ main() {
   printf "\n"
   printf "  ${bold}Quick start${reset}\n"
   printf "\n"
-  printf "    ${green}termpair share${reset}                  Private encrypted session\n"
-  printf "    ${green}termpair share --public${reset}         Public, read-only, no encryption\n"
-  printf "    ${green}termpair share --read-only${reset}      Viewers can watch but not type\n"
-  printf "    ${green}termpair serve --port 8000${reset}      Run your own server\n"
+  printf "    ${green}termpair${reset}                          Private encrypted session\n"
+  printf "    ${green}termpair --public${reset}                 Public, read-only, no encryption\n"
+  printf "    ${green}termpair --read-only${reset}              Viewers can watch but not type\n"
+  printf "    ${green}termpair --host https://my-server${reset} Share via a custom server\n"
   printf "\n"
   printf "  ${dim}https://github.com/cs01/termpair${reset}\n"
   printf "\n"
