+++ termpair-rs/crates/termpair-common/Cargo.toml
@@ -1,0 +1,18 @@
+[package]
+name = "termpair-common"
+version = "1.2.0"
+edition = "2021"
+license = "MIT"
+
+[dependencies]
+aes-gcm = "0.10"
+hmac = "0.12"
+sha2 = "0.10"
+hex = "0.4"
+zeroize = "1"
+subtle = "2"
+rand = "0.8"
+base64 = "0.22"
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+chrono = "0.4"
