+++ termpair-rs/crates/termpair-common/src/constants.rs
