+++ termpair-rs/crates/termpair-common/src/encryption.rs
