+++ termpair-rs/crates/termpair-common/src/types.rs
