+++ termpair-rs/crates/termpair-server/Cargo.toml
@@ -1,0 +1,33 @@
+[package]
+name = "termpair-server"
+version = "1.2.0"
+edition = "2021"
+description = "TermPair server — routes encrypted terminal sharing sessions"
+license = "MIT"
+
+[[bin]]
+name = "termpair-server"
+path = "src/main.rs"
+
+[dependencies]
+termpair-common = { path = "../termpair-common" }
+clap = { version = "4", features = ["derive"] }
+axum = { version = "0.8", features = ["ws", "query"] }
+tokio = { version = "1", features = ["full"] }
+serde = { version = "1", features = ["derive"] }
+serde_json = "1"
+futures-util = "0.3"
+base64 = "0.22"
+hmac = "0.12"
+sha2 = "0.10"
+hex = "0.4"
+subtle = "2"
+rand = "0.8"
+tracing = "0.1"
+tracing-subscriber = "0.3"
+tower-http = { version = "0.6", features = ["cors", "fs", "set-header"] }
+rust-embed = { version = "8", features = ["compression"] }
+mime_guess = "2"
+axum-server = { version = "0.7", features = ["tls-rustls"] }
+chrono = "0.4"
+dirs = "6"
