+++ termpair-rs/crates/termpair-server/src/lib.rs
@@ -1,0 +1,13 @@
+pub mod names;
+pub mod server;
+
+use rand::rngs::OsRng;
+use rand::Rng;
+
+pub fn random_string(n: usize) -> String {
+    const CHARS: &[u8] = b"abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789";
+    let mut rng = OsRng;
+    (0..n)
+        .map(|_| CHARS[rng.gen_range(0..CHARS.len())] as char)
+        .collect()
+}
