+++ termpair-rs/crates/termpair/src/share/mod.rs
